+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.0]
 ### Added
 - Added `rcc::Lsco` to use the low-speed oscillator output.
 - Added `Rtc::calibrate_lp` to calibrate the RTC.
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Rtc::is_alarm_{a,b}_en`.
 - Added methods to utilize ADC oversampling.
 
-## Changed
+### Changed
 - `Rtc.alarm_{a,b}` returns `Alarm` instead of `Option<Alarm>`.
 - Updated `stm32-rs` from `0.14.0` to `0.15.1`.
 
@@ -115,7 +115,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [tweedegolf]: https://github.com/tweedegolf
 [stm32-rs]: https://github.com/stm32-rs
-[Unreleased]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.5.1...HEAD
+[Unreleased]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.4.0...v0.4.1